requirements.txt
discord.py==1.2.5
google==2.0.2
python-dotenv==0.10.3
mysql-connector-python==8.0.18